requirements.txt
fastapi
uvicorn
icalendar
requests
python-dotenv
pytz